refactor(runtime): split ops.rs into cohesive submodules (§7-8)
Split the 1721-line runtime/ops.rs (over the 500-line limit) into four
cohesive modules under src/runtime/. Three are well under 500 lines;
ops_reduction.rs (829) holds the single ~717-line apply_reduction_op
method, which cannot be split at a function boundary. All methods are
impl Interpreter, so the split is visibility-only.

- ops_set.rs (585)       - Set/Bag/Mix union, equality, multiply, addition ops
- ops_reduction.rs (829) - reduction helpers + apply_reduction_op (dominant method)
- ops_compare.rs (168)   - value_to_list + compare
- ops_bits.rs (148)      - buf/string bitwise ops + integer shift ops

Pure relocation. Nine private methods bumped to pub(crate) for
cross-file calls from apply_reduction_op: apply_set_union,
apply_set_equality, apply_set_multiply, apply_set_addition,
str_bitwise_op, shift_left_i64, shift_right_i64, shift_left_bigint,
shift_right_bigint. The fn multiset is byte-identical. No behavior change.
