+++ src/runtime/mod.rs
@@ -267,7 +267,10 @@ mod native_supply_methods;
 mod native_supply_mut_methods;
 pub(crate) mod native_types;
 pub(crate) mod nativecall;
-mod ops;
+mod ops_bits;
+mod ops_compare;
+mod ops_reduction;
+mod ops_set;
 mod output_sink;
 pub(crate) mod phasers;
 mod react_died;
