+++ src/runtime/ops_reduction.rs
@@ -3,588 +3,6 @@ use crate::symbol::Symbol;
 use num_traits::{Signed, ToPrimitive, Zero};
 
 impl Interpreter {
-    fn union_is_infinite_bound(value: &Value) -> bool {
-        match value {
-            Value::Num(n) => n.is_infinite(),
-            Value::Rat(_, d) | Value::FatRat(_, d) => *d == 0,
-            Value::Mixin(inner, _) => Self::union_is_infinite_bound(inner),
-            _ => false,
-        }
-    }
-
-    fn union_is_lazy_input(value: &Value) -> bool {
-        match value {
-            Value::LazyList(_) => true,
-            Value::GenericRange { start, end, .. } => {
-                Self::union_is_infinite_bound(start) || Self::union_is_infinite_bound(end)
-            }
-            _ => false,
-        }
-    }
-
-    fn union_set_keys(value: &Value) -> Result<std::collections::HashSet<String>, RuntimeError> {
-        if Self::union_is_lazy_input(value) {
-            return Err(RuntimeError::new("X::Cannot::Lazy"));
-        }
-        match value {
-            Value::Set(s, _) => Ok(s.elements.clone()),
-            Value::Bag(b, _) => Ok(b.keys().cloned().collect()),
-            Value::Mix(m, _) => Ok(m.keys().cloned().collect()),
-            Value::Hash(h) => Ok(h
-                .iter()
-                .filter_map(|(k, v)| {
-                    if v.truthy() || matches!(v, Value::Nil) {
-                        Some(k.clone())
-                    } else {
-                        None
-                    }
-                })
-                .collect()),
-            _ if value.as_list_items().is_some() => {
-                let mut elems = std::collections::HashSet::new();
-                for item in value.as_list_items().unwrap().iter() {
-                    Self::union_insert_set_elem(&mut elems, item);
-                }
-                Ok(elems)
-            }
-            range if range.is_range() => {
-                let mut elems = std::collections::HashSet::new();
-                for item in Self::value_to_list(range) {
-                    Self::union_insert_set_elem(&mut elems, &item);
-                }
-                Ok(elems)
-            }
-            Value::Pair(_, _) | Value::ValuePair(_, _) => {
-                let mut elems = std::collections::HashSet::new();
-                Self::union_insert_set_elem(&mut elems, value);
-                Ok(elems)
-            }
-            other => {
-                let mut elems = std::collections::HashSet::new();
-                let sv = other.to_string_value();
-                if !sv.is_empty() {
-                    elems.insert(sv);
-                }
-                Ok(elems)
-            }
-        }
-    }
-
-    fn union_bag_counts(
-        value: &Value,
-    ) -> Result<std::collections::HashMap<String, i64>, RuntimeError> {
-        if Self::union_is_lazy_input(value) {
-            return Err(RuntimeError::new("X::Cannot::Lazy"));
-        }
-        match value {
-            Value::Bag(b, _) => Ok(crate::runtime::utils::bag_counts_as_i64(&b.counts)),
-            Value::Mix(m, _) => Ok(m
-                .iter()
-                .filter_map(|(k, w)| {
-                    if *w != 0.0 {
-                        Some((k.clone(), 1))
-                    } else {
-                        None
-                    }
-                })
-                .collect()),
-            other => {
-                let set = Self::union_set_keys(other)?;
-                Ok(set.into_iter().map(|k| (k, 1)).collect())
-            }
-        }
-    }
-
-    fn union_mix_weights(
-        value: &Value,
-    ) -> Result<std::collections::HashMap<String, f64>, RuntimeError> {
-        if Self::union_is_lazy_input(value) {
-            return Err(RuntimeError::new("X::Cannot::Lazy"));
-        }
-        match value {
-            Value::Mix(m, _) => Ok(m.weights.clone()),
-            Value::Bag(b, _) => Ok(b
-                .iter()
-                .map(|(k, v)| (k.clone(), crate::runtime::utils::bigint_to_f64_sat(v)))
-                .collect()),
-            other => {
-                let set = Self::union_set_keys(other)?;
-                Ok(set.into_iter().map(|k| (k, 1.0)).collect())
-            }
-        }
-    }
-
-    fn apply_set_union(left: &Value, right: &Value) -> Result<Value, RuntimeError> {
-        if matches!(left, Value::Instance { class_name, .. } if class_name == "Failure")
-            || matches!(right, Value::Instance { class_name, .. } if class_name == "Failure")
-        {
-            return Err(RuntimeError::new("Exception"));
-        }
-        if matches!(left, Value::Mix(_, _)) || matches!(right, Value::Mix(_, _)) {
-            let mut l = Self::union_mix_weights(left)?;
-            let r = Self::union_mix_weights(right)?;
-            for (k, v) in r {
-                let e = l.entry(k).or_insert(0.0);
-                *e = e.max(v);
-            }
-            return Ok(Value::mix(l));
-        }
-        if matches!(left, Value::Bag(_, _)) || matches!(right, Value::Bag(_, _)) {
-            let mut l = Self::union_bag_counts(left)?;
-            let r = Self::union_bag_counts(right)?;
-            for (k, v) in r {
-                let e = l.entry(k).or_insert(0);
-                *e = (*e).max(v);
-            }
-            return Ok(Value::bag(l));
-        }
-        let mut l = Self::union_set_keys(left)?;
-        let r = Self::union_set_keys(right)?;
-        l.extend(r);
-        Ok(Value::set(l))
-    }
-
-    fn set_equal_bag_counts(
-        value: &Value,
-    ) -> Result<std::collections::HashMap<String, i64>, RuntimeError> {
-        let mut counts = Self::union_bag_counts(value)?;
-        counts.retain(|_, v| *v > 0);
-        Ok(counts)
-    }
-
-    fn set_equal_mix_weights(
-        value: &Value,
-    ) -> Result<std::collections::HashMap<String, f64>, RuntimeError> {
-        let mut weights = Self::union_mix_weights(value)?;
-        weights.retain(|_, w| *w != 0.0);
-        Ok(weights)
-    }
-
-    fn apply_set_equality(left: &Value, right: &Value) -> Result<bool, RuntimeError> {
-        if matches!(left, Value::Mix(_, _)) || matches!(right, Value::Mix(_, _)) {
-            return Ok(Self::set_equal_mix_weights(left)? == Self::set_equal_mix_weights(right)?);
-        }
-        if matches!(left, Value::Bag(_, _)) || matches!(right, Value::Bag(_, _)) {
-            return Ok(Self::set_equal_bag_counts(left)? == Self::set_equal_bag_counts(right)?);
-        }
-        Ok(Self::union_set_keys(left)? == Self::union_set_keys(right)?)
-    }
-
-    fn multiply_pair_i64(value: &Value) -> i64 {
-        match value {
-            Value::Int(i) => *i,
-            Value::Num(n) => *n as i64,
-            Value::Rat(n, d) if *d != 0 => n / d,
-            Value::FatRat(n, d) if *d != 0 => n / d,
-            Value::Bool(b) => i64::from(*b),
-            _ => i64::from(value.truthy()),
-        }
-    }
-
-    fn multiply_pair_f64(value: &Value) -> f64 {
-        match value {
-            Value::Int(i) => *i as f64,
-            Value::Num(n) => *n,
-            Value::Rat(n, d) if *d != 0 => *n as f64 / *d as f64,
-            Value::FatRat(n, d) if *d != 0 => *n as f64 / *d as f64,
-            _ => 1.0,
-        }
-    }
-
-    fn multiply_bag_counts(
-        value: &Value,
-    ) -> Result<std::collections::HashMap<String, (i64, bool)>, RuntimeError> {
-        let value = value.descalarize();
-        if Self::union_is_lazy_input(value) {
-            return Err(RuntimeError::new("X::Cannot::Lazy"));
-        }
-        match value {
-            Value::Bag(b, _) => {
-                let resolved = crate::runtime::utils::resolve_bag_tab_keys(b);
-                Ok(resolved.into_iter().map(|(k, v)| (k, (v, false))).collect())
-            }
-            Value::Mix(m, _) => Ok(m
-                .iter()
-                .map(|(k, v)| (k.clone(), (*v as i64, false)))
-                .collect()),
-            Value::Set(s, _) => Ok(s.iter().map(|k| (k.clone(), (1, false))).collect()),
-            Value::Hash(h) => Ok(h
-                .iter()
-                .filter_map(|(k, v)| {
-                    let c = Self::multiply_pair_i64(v);
-                    if c > 0 {
-                        Some((k.clone(), (c, c != 1)))
-                    } else {
-                        None
-                    }
-                })
-                .collect()),
-            _ if value.as_list_items().is_some() => {
-                let mut counts = std::collections::HashMap::new();
-                for item in value.as_list_items().unwrap().iter() {
-                    match item {
-                        Value::Pair(k, v) => {
-                            let c = Self::multiply_pair_i64(v);
-                            let entry = counts.entry(k.clone()).or_insert((0, false));
-                            entry.0 += c;
-                            entry.1 |= c != 1;
-                        }
-                        Value::ValuePair(k, v) => {
-                            let key = k.to_string_value();
-                            let c = Self::multiply_pair_i64(v);
-                            let entry = counts.entry(key).or_insert((0, false));
-                            entry.0 += c;
-                            entry.1 |= c != 1;
-                        }
-                        other => {
-                            let entry = counts.entry(other.to_string_value()).or_insert((0, false));
-                            entry.0 += 1;
-                        }
-                    }
-                }
-                Ok(counts)
-            }
-            range if range.is_range() => {
-                let mut counts = std::collections::HashMap::new();
-                for item in Self::value_to_list(range) {
-                    let entry = counts.entry(item.to_string_value()).or_insert((0, false));
-                    entry.0 += 1;
-                }
-                Ok(counts)
-            }
-            Value::Pair(k, v) => {
-                let mut counts = std::collections::HashMap::new();
-                let c = Self::multiply_pair_i64(v);
-                if c > 0 {
-                    counts.insert(k.clone(), (c, c != 1));
-                }
-                Ok(counts)
-            }
-            Value::ValuePair(k, v) => {
-                let mut counts = std::collections::HashMap::new();
-                let c = Self::multiply_pair_i64(v);
-                if c > 0 {
-                    counts.insert(k.to_string_value(), (c, c != 1));
-                }
-                Ok(counts)
-            }
-            other => {
-                let mut counts = std::collections::HashMap::new();
-                counts.insert(other.to_string_value(), (1, false));
-                Ok(counts)
-            }
-        }
-    }
-
-    fn multiply_mix_weights(
-        value: &Value,
-    ) -> Result<std::collections::HashMap<String, f64>, RuntimeError> {
-        let value = value.descalarize();
-        if Self::union_is_lazy_input(value) {
-            return Err(RuntimeError::new("X::Cannot::Lazy"));
-        }
-        match value {
-            Value::Mix(m, _) => Ok(m.weights.clone()),
-            Value::Bag(b, _) => {
-                let resolved = crate::runtime::utils::resolve_bag_tab_keys(b);
-                Ok(resolved.into_iter().map(|(k, v)| (k, v as f64)).collect())
-            }
-            Value::Set(s, _) => Ok(s.iter().map(|k| (k.clone(), 1.0)).collect()),
-            Value::Hash(h) => Ok(h
-                .iter()
-                .filter_map(|(k, v)| {
-                    let w = match v {
-                        Value::Int(i) => *i as f64,
-                        Value::Num(n) => *n,
-                        Value::Rat(n, d) if *d != 0 => *n as f64 / *d as f64,
-                        Value::FatRat(n, d) if *d != 0 => *n as f64 / *d as f64,
-                        Value::Bool(b) => {
-                            if *b {
-                                1.0
-                            } else {
-                                0.0
-                            }
-                        }
-                        _ => {
-                            if v.truthy() {
-                                1.0
-                            } else {
-                                0.0
-                            }
-                        }
-                    };
-                    if w != 0.0 { Some((k.clone(), w)) } else { None }
-                })
-                .collect()),
-            _ if value.as_list_items().is_some() => {
-                let mut weights = std::collections::HashMap::new();
-                for item in value.as_list_items().unwrap().iter() {
-                    match item {
-                        Value::Pair(k, v) => {
-                            *weights.entry(k.clone()).or_insert(0.0) += Self::multiply_pair_f64(v);
-                        }
-                        Value::ValuePair(k, v) => {
-                            let key = k.to_string_value();
-                            *weights.entry(key).or_insert(0.0) += Self::multiply_pair_f64(v);
-                        }
-                        other => {
-                            *weights.entry(other.to_string_value()).or_insert(0.0) += 1.0;
-                        }
-                    }
-                }
-                Ok(weights)
-            }
-            range if range.is_range() => {
-                let mut weights = std::collections::HashMap::new();
-                for item in Self::value_to_list(range) {
-                    *weights.entry(item.to_string_value()).or_insert(0.0) += 1.0;
-                }
-                Ok(weights)
-            }
-            Value::Pair(k, v) => {
-                let mut weights = std::collections::HashMap::new();
-                let w = Self::multiply_pair_f64(v);
-                if w != 0.0 {
-                    weights.insert(k.clone(), w);
-                }
-                Ok(weights)
-            }
-            Value::ValuePair(k, v) => {
-                let mut weights = std::collections::HashMap::new();
-                let w = Self::multiply_pair_f64(v);
-                if w != 0.0 {
-                    weights.insert(k.to_string_value(), w);
-                }
-                Ok(weights)
-            }
-            other => {
-                let mut weights = std::collections::HashMap::new();
-                weights.insert(other.to_string_value(), 1.0);
-                Ok(weights)
-            }
-        }
-    }
-
-    fn apply_set_multiply(left: &Value, right: &Value) -> Result<Value, RuntimeError> {
-        let left = match left {
-            Value::Scalar(inner) => inner.as_ref(),
-            other => other,
-        };
-        let right = match right {
-            Value::Scalar(inner) => inner.as_ref(),
-            other => other,
-        };
-        if matches!(left, Value::Instance { class_name, .. } if class_name == "Failure")
-            || matches!(right, Value::Instance { class_name, .. } if class_name == "Failure")
-        {
-            return Err(RuntimeError::new("Exception"));
-        }
-        if matches!(left, Value::Mix(_, _)) || matches!(right, Value::Mix(_, _)) {
-            let l = Self::multiply_mix_weights(left)?;
-            let r = Self::multiply_mix_weights(right)?;
-            let mut result = std::collections::HashMap::new();
-            for (k, lv) in l {
-                if let Some(rv) = r.get(&k) {
-                    let product = lv * rv;
-                    if product != 0.0 {
-                        result.insert(k, product);
-                    }
-                }
-            }
-            return Ok(Value::mix(result));
-        }
-        let l = Self::multiply_bag_counts(left)?;
-        let r = Self::multiply_bag_counts(right)?;
-        let mut result: std::collections::HashMap<String, i64> = std::collections::HashMap::new();
-        for (k, (lv, _l_explicit)) in l {
-            if let Some((rv, _r_explicit)) = r.get(&k) {
-                let product = lv * *rv;
-                if product > 0 {
-                    result.insert(k, product);
-                }
-            }
-        }
-        Ok(Value::bag(result))
-    }
-
-    /// Coerce a value to bag-like count map for the (+) operator.
-    /// Unlike union_bag_counts, this properly handles Hash and Pair values.
-    fn addition_bag_counts(
-        value: &Value,
-    ) -> Result<std::collections::HashMap<String, i64>, RuntimeError> {
-        if Self::union_is_lazy_input(value) {
-            return Err(RuntimeError::new("X::Cannot::Lazy"));
-        }
-        match value {
-            Value::Bag(b, _) => Ok(crate::runtime::utils::bag_counts_as_i64(&b.counts)),
-            Value::Mix(m, _) => Ok(m
-                .iter()
-                .filter_map(|(k, w)| {
-                    if *w != 0.0 {
-                        Some((k.clone(), *w as i64))
-                    } else {
-                        None
-                    }
-                })
-                .collect()),
-            Value::Set(s, _) => Ok(s.iter().map(|k| (k.clone(), 1)).collect()),
-            Value::Hash(map) => {
-                let mut result = std::collections::HashMap::new();
-                for (k, v) in map.iter() {
-                    let weight = v.to_f64() as i64;
-                    if weight != 0 {
-                        result.insert(k.clone(), weight);
-                    }
-                }
-                Ok(result)
-            }
-            _ if value.as_list_items().is_some() => {
-                let mut result = std::collections::HashMap::new();
-                for item in value.as_list_items().unwrap().iter() {
-                    match item {
-                        Value::Pair(k, v) => {
-                            let weight = v.to_f64() as i64;
-                            *result.entry(k.clone()).or_insert(0) += weight;
-                        }
-                        Value::ValuePair(k, v) => {
-                            let weight = v.to_f64() as i64;
-                            *result.entry(k.to_string_value()).or_insert(0) += weight;
-                        }
-                        other => {
-                            let key = other.to_string_value();
-                            if !key.is_empty() {
-                                *result.entry(key).or_insert(0) += 1;
-                            }
-                        }
-                    }
-                }
-                result.retain(|_, v| *v != 0);
-                Ok(result)
-            }
-            Value::Pair(k, v) => {
-                let mut result = std::collections::HashMap::new();
-                let weight = v.to_f64() as i64;
-                if weight != 0 {
-                    result.insert(k.clone(), weight);
-                }
-                Ok(result)
-            }
-            other => {
-                let set = Self::union_set_keys(other)?;
-                Ok(set.into_iter().map(|k| (k, 1)).collect())
-            }
-        }
-    }
-
-    /// Coerce a value to mix-like weights for the (+) operator.
-    fn addition_mix_weights(
-        value: &Value,
-    ) -> Result<std::collections::HashMap<String, f64>, RuntimeError> {
-        if Self::union_is_lazy_input(value) {
-            return Err(RuntimeError::new("X::Cannot::Lazy"));
-        }
-        match value {
-            Value::Mix(m, _) => Ok(m.weights.clone()),
-            Value::Bag(b, _) => Ok(b
-                .iter()
-                .map(|(k, v)| (k.clone(), crate::runtime::utils::bigint_to_f64_sat(v)))
-                .collect()),
-            Value::Set(s, _) => Ok(s.iter().map(|k| (k.clone(), 1.0)).collect()),
-            Value::Hash(map) => {
-                let mut result = std::collections::HashMap::new();
-                for (k, v) in map.iter() {
-                    let weight = v.to_f64();
-                    if weight != 0.0 {
-                        result.insert(k.clone(), weight);
-                    }
-                }
-                Ok(result)
-            }
-            _ if value.as_list_items().is_some() => {
-                let mut result = std::collections::HashMap::new();
-                for item in value.as_list_items().unwrap().iter() {
-                    match item {
-                        Value::Pair(k, v) => {
-                            let weight = v.to_f64();
-                            *result.entry(k.clone()).or_insert(0.0) += weight;
-                        }
-                        Value::ValuePair(k, v) => {
-                            let weight = v.to_f64();
-                            *result.entry(k.to_string_value()).or_insert(0.0) += weight;
-                        }
-                        other => {
-                            let key = other.to_string_value();
-                            if !key.is_empty() {
-                                *result.entry(key).or_insert(0.0) += 1.0;
-                            }
-                        }
-                    }
-                }
-                result.retain(|_, v| *v != 0.0);
-                Ok(result)
-            }
-            Value::Pair(k, v) => {
-                let mut result = std::collections::HashMap::new();
-                let weight = v.to_f64();
-                if weight != 0.0 {
-                    result.insert(k.clone(), weight);
-                }
-                Ok(result)
-            }
-            other => {
-                let set = Self::union_set_keys(other)?;
-                Ok(set.into_iter().map(|k| (k, 1.0)).collect())
-            }
-        }
-    }
-
-    fn apply_set_addition(left: &Value, right: &Value) -> Result<Value, RuntimeError> {
-        let left = match left {
-            Value::Scalar(inner) => inner.as_ref(),
-            other => other,
-        };
-        let right = match right {
-            Value::Scalar(inner) => inner.as_ref(),
-            other => other,
-        };
-        if matches!(left, Value::Instance { class_name, .. } if class_name == "Failure")
-            || matches!(right, Value::Instance { class_name, .. } if class_name == "Failure")
-        {
-            return Err(RuntimeError::new("Exception"));
-        }
-        // Determine type level: Mix > Bag > Set, minimum is Bag for (+)
-        let type_level = |v: &Value| -> u8 {
-            match v {
-                Value::Mix(_, _) => 2,
-                Value::Bag(_, _) => 1,
-                Value::Package(sym) => match sym.resolve().as_str() {
-                    "Mix" | "MixHash" => 2,
-                    "Bag" | "BagHash" => 1,
-                    _ => 0,
-                },
-                _ => 0,
-            }
-        };
-        let result_level = type_level(left).max(type_level(right)).max(1);
-
-        if result_level >= 2 {
-            let mut l = Self::addition_mix_weights(left)?;
-            let r = Self::addition_mix_weights(right)?;
-            for (k, v) in r {
-                let e = l.entry(k).or_insert(0.0);
-                *e += v;
-            }
-            return Ok(Value::mix(l));
-        }
-        let mut l = Self::addition_bag_counts(left)?;
-        let r = Self::addition_bag_counts(right)?;
-        for (k, v) in r {
-            let e = l.entry(k).or_insert(0);
-            *e += v;
-        }
-        Ok(Value::bag(l))
-    }
-
     fn reduction_repeat_error(class_name: &str, message: &str) -> RuntimeError {
         let mut attrs = std::collections::HashMap::new();
         attrs.insert("message".to_string(), Value::str(message.to_string()));
@@ -689,151 +107,6 @@ impl Interpreter {
             )),
             Value::Package(_) => Ok(Some(0)),
             _ => Ok(Some(0)),
-        }
-    }
-
-    fn buf_class_name(val: &Value) -> Option<String> {
-        if let Value::Instance { class_name, .. } = val {
-            let cn = class_name.resolve();
-            if cn == "Buf"
-                || cn == "Blob"
-                || cn == "utf8"
-                || cn == "utf16"
-                || cn.starts_with("Buf[")
-                || cn.starts_with("Blob[")
-                || cn.starts_with("buf")
-                || cn.starts_with("blob")
-            {
-                return Some(cn.to_string());
-            }
-        }
-        None
-    }
-
-    fn str_bitwise_op(
-        left: &Value,
-        right: &Value,
-        op: fn(u32, u32) -> u32,
-        pad_to_max: bool,
-    ) -> Result<Value, RuntimeError> {
-        let left_is_buf = Self::is_buf_value(left);
-        let right_is_buf = Self::is_buf_value(right);
-        let any_buf = left_is_buf || right_is_buf;
-        if any_buf {
-            // For Buf values, operate on bytes
-            let lb = if left_is_buf {
-                Self::extract_buf_bytes(left)
-            } else {
-                crate::runtime::utils::coerce_to_str(left)
-                    .as_bytes()
-                    .to_vec()
-            };
-            let rb = if right_is_buf {
-                Self::extract_buf_bytes(right)
-            } else {
-                crate::runtime::utils::coerce_to_str(right)
-                    .as_bytes()
-                    .to_vec()
-            };
-            // Buf bitwise ops always extend to the longer operand,
-            // padding the shorter one with zeros (even for ~&).
-            let len = lb.len().max(rb.len());
-            let mut out = Vec::with_capacity(len);
-            for i in 0..len {
-                let a = lb.get(i).copied().unwrap_or(0) as u32;
-                let b = rb.get(i).copied().unwrap_or(0) as u32;
-                out.push(op(a, b) as u8);
-            }
-            let byte_vals: Vec<Value> = out.into_iter().map(|b| Value::Int(b as i64)).collect();
-            let mut attrs = std::collections::HashMap::new();
-            attrs.insert("bytes".to_string(), Value::array(byte_vals));
-            let result_type = match (Self::buf_class_name(left), Self::buf_class_name(right)) {
-                (Some(l), Some(r)) if l == r => l,
-                _ => "Buf".to_string(),
-            };
-            Ok(Value::make_instance(
-                crate::symbol::Symbol::intern(&result_type),
-                attrs,
-            ))
-        } else {
-            // For strings, operate on Unicode codepoints (ordinal values)
-            let ls = crate::runtime::utils::coerce_to_str(left);
-            let rs = crate::runtime::utils::coerce_to_str(right);
-            let lc: Vec<u32> = ls.chars().map(|c| c as u32).collect();
-            let rc: Vec<u32> = rs.chars().map(|c| c as u32).collect();
-            let len = if pad_to_max {
-                lc.len().max(rc.len())
-            } else {
-                lc.len().min(rc.len())
-            };
-            let mut out = String::with_capacity(len);
-            for i in 0..len {
-                let a = lc.get(i).copied().unwrap_or(0);
-                let b = rc.get(i).copied().unwrap_or(0);
-                let result_cp = op(a, b);
-                if let Some(ch) = char::from_u32(result_cp) {
-                    out.push(ch);
-                }
-            }
-            // Apply NFC normalization to the result
-            use unicode_normalization::UnicodeNormalization;
-            Ok(Value::str(out.nfc().collect::<String>()))
-        }
-    }
-
-    fn shift_left_i64(a: i64, b: i64) -> Value {
-        if b < 0 {
-            let shift = b.unsigned_abs();
-            let shifted = if shift >= i64::BITS as u64 {
-                if a < 0 { -1 } else { 0 }
-            } else {
-                a >> (shift as u32)
-            };
-            return Value::Int(shifted);
-        }
-        let shift = b as u64;
-        if shift >= i64::BITS as u64 {
-            return Value::from_bigint(num_bigint::BigInt::from(a) << (shift as usize));
-        }
-        // Use BigInt for the shift to avoid i64 overflow (Raku integers are arbitrary precision)
-        Value::from_bigint(num_bigint::BigInt::from(a) << (shift as usize))
-    }
-
-    fn shift_right_i64(a: i64, b: i64) -> Value {
-        if b < 0 {
-            let shift = b.unsigned_abs();
-            if shift >= i64::BITS as u64 {
-                return Value::from_bigint(num_bigint::BigInt::from(a) << (shift as usize));
-            }
-            if let Some(v) = a.checked_shl(shift as u32) {
-                Value::Int(v)
-            } else {
-                Value::from_bigint(num_bigint::BigInt::from(a) << (shift as usize))
-            }
-        } else {
-            let shift = b as u64;
-            let shifted = if shift >= i64::BITS as u64 {
-                if a < 0 { -1 } else { 0 }
-            } else {
-                a >> (shift as u32)
-            };
-            Value::Int(shifted)
-        }
-    }
-
-    fn shift_left_bigint(a: &num_bigint::BigInt, b: i64) -> Value {
-        if b < 0 {
-            Value::from_bigint(a >> (b.unsigned_abs() as usize))
-        } else {
-            Value::from_bigint(a << (b as usize))
-        }
-    }
-
-    fn shift_right_bigint(a: &num_bigint::BigInt, b: i64) -> Value {
-        if b < 0 {
-            Value::from_bigint(a << (b.unsigned_abs() as usize))
-        } else {
-            Value::from_bigint(a >> (b as usize))
         }
     }
 
@@ -1551,171 +824,6 @@ impl Interpreter {
                 "Unsupported reduction operator: {}",
                 op
             ))),
-        }
-    }
-
-    pub(crate) fn value_to_list(val: &Value) -> Vec<Value> {
-        match val {
-            Value::Array(items, kind) if kind.is_itemized() => vec![val.clone()],
-            Value::Array(items, ..) => items.to_vec(),
-            Value::Seq(items) => items.to_vec(),
-            Value::LazyList(ll) => ll.cache.lock().unwrap().clone().unwrap_or_default(),
-            Value::Hash(items) => items
-                .iter()
-                .map(|(k, v)| items.typed_pair(k, v.clone()))
-                .collect(),
-            Value::Range(a, b) => {
-                let end = (*b).min(*a + 1_000_000);
-                (*a..=end).map(Value::Int).collect()
-            }
-            Value::RangeExcl(a, b) => {
-                let end = (*b).min(*a + 1_000_000);
-                (*a..end).map(Value::Int).collect()
-            }
-            Value::RangeExclStart(a, b) => {
-                let start = *a + 1;
-                let end = (*b).min(start + 1_000_000);
-                (start..=end).map(Value::Int).collect()
-            }
-            Value::RangeExclBoth(a, b) => {
-                let start = *a + 1;
-                let end = (*b).min(start + 1_000_000);
-                (start..end).map(Value::Int).collect()
-            }
-            Value::GenericRange { .. } => crate::runtime::utils::value_to_list(val),
-            Value::Set(items, _) => items
-                .iter()
-                .map(|s| Value::Pair(s.clone(), Box::new(Value::Bool(true))))
-                .collect(),
-            Value::Bag(items, _) => items
-                .iter()
-                .map(|(k, v)| Value::Pair(k.clone(), Box::new(Value::from_bigint(v.clone()))))
-                .collect(),
-            Value::Mix(items, _) => items
-                .iter()
-                .map(|(k, v)| {
-                    Value::Pair(k.clone(), Box::new(crate::value::mix_weight_to_value(*v)))
-                })
-                .collect(),
-            Value::Slip(items) => items.to_vec(),
-            // A WalkList flattens to its candidate closures in list context.
-            Value::Instance {
-                class_name,
-                attributes,
-                ..
-            } if class_name.resolve() == "WalkList" => {
-                crate::runtime::utils::walk_list_candidates(attributes)
-                    .unwrap_or_else(|| vec![val.clone()])
-            }
-            // Nil is a single scalar item in list context (e.g. `for Nil { }`
-            // does one iteration); it is not an empty list.
-            other => vec![other.clone()],
-        }
-    }
-
-    pub(crate) fn compare(
-        left: Value,
-        right: Value,
-        f: fn(i32) -> bool,
-    ) -> Result<Value, RuntimeError> {
-        if matches!(left, Value::Pair(..) | Value::ValuePair(..))
-            || matches!(right, Value::Pair(..) | Value::ValuePair(..))
-        {
-            return Err(RuntimeError::typed_msg(
-                "X::Multi::NoMatch",
-                "Cannot resolve caller; none of the candidates match",
-            ));
-        }
-        // Version-vs-Version comparison: use version_cmp_parts directly
-        if let (Value::Version { parts: ap, .. }, Value::Version { parts: bp, .. }) =
-            (&left, &right)
-        {
-            let ord = super::version_cmp_parts(ap, bp) as i32;
-            return Ok(Value::Bool(f(ord)));
-        }
-        let (l, r) = super::coerce_numeric(left, right);
-        // Complex numbers cannot be ordered; throw if either operand has non-zero imaginary part
-        if let Value::Complex(_, im) = &l
-            && *im != 0.0
-        {
-            return Err(RuntimeError::new(
-                "Cannot convert Complex to Real: imaginary part not zero",
-            ));
-        }
-        if let Value::Complex(_, im) = &r
-            && *im != 0.0
-        {
-            return Err(RuntimeError::new(
-                "Cannot convert Complex to Real: imaginary part not zero",
-            ));
-        }
-        if let (Some(a), Some(b)) = (super::to_big_rat_parts(&l), super::to_big_rat_parts(&r))
-            && (matches!(
-                l,
-                Value::Rat(_, _) | Value::FatRat(_, _) | Value::BigRat(_, _)
-            ) || matches!(
-                r,
-                Value::Rat(_, _) | Value::FatRat(_, _) | Value::BigRat(_, _)
-            ))
-        {
-            if let Some(ord) = super::compare_big_rat_parts(a, b) {
-                return Ok(Value::Bool(f(ord as i32)));
-            }
-            return Ok(Value::Bool(false));
-        }
-        match (l, r) {
-            (Value::Int(a), Value::Int(b)) => {
-                let ord = a.cmp(&b) as i32;
-                Ok(Value::Bool(f(ord)))
-            }
-            (Value::Num(a), Value::Num(b)) => {
-                // NaN is unordered: all comparisons except != return False
-                if let Some(ord) = a.partial_cmp(&b) {
-                    Ok(Value::Bool(f(ord as i32)))
-                } else {
-                    Ok(Value::Bool(false))
-                }
-            }
-            (Value::Int(a), Value::Num(b)) => {
-                let a = a as f64;
-                if let Some(ord) = a.partial_cmp(&b) {
-                    Ok(Value::Bool(f(ord as i32)))
-                } else {
-                    Ok(Value::Bool(false))
-                }
-            }
-            (Value::Num(a), Value::Int(b)) => {
-                let b = b as f64;
-                if let Some(ord) = a.partial_cmp(&b) {
-                    Ok(Value::Bool(f(ord as i32)))
-                } else {
-                    Ok(Value::Bool(false))
-                }
-            }
-            // BigInt vs BigInt
-            (Value::BigInt(a), Value::BigInt(b)) => {
-                let ord = a.cmp(&b) as i32;
-                Ok(Value::Bool(f(ord)))
-            }
-            // BigInt vs Int (and vice versa) — exact comparison
-            (Value::BigInt(a), Value::Int(b)) => {
-                let ord = a.as_ref().cmp(&num_bigint::BigInt::from(b)) as i32;
-                Ok(Value::Bool(f(ord)))
-            }
-            (Value::Int(a), Value::BigInt(b)) => {
-                let ord = num_bigint::BigInt::from(a).cmp(b.as_ref()) as i32;
-                Ok(Value::Bool(f(ord)))
-            }
-            (ref left_val, ref right_val) => {
-                // Fallback: convert both to f64 for cross-type comparisons (e.g. Num vs Rat)
-                let a = super::to_float_value(left_val).unwrap_or(0.0);
-                let b = super::to_float_value(right_val).unwrap_or(0.0);
-                if let Some(ord) = a.partial_cmp(&b) {
-                    Ok(Value::Bool(f(ord as i32)))
-                } else {
-                    Ok(Value::Bool(false))
-                }
-            }
         }
     }
 }
